test: rearrange n separate tests
